+++ Sources/Sentry/SentryCrashIntegration.m
@@ -233,6 +233,8 @@ SentryCrashIntegration ()
 
 #if TARGET_OS_SIMULATOR
     [deviceData setValue:@(YES) forKey:@"simulator"];
+#else
+    [deviceData setValue:@(NO) forKey:@"simulator"];
 #endif
 
     NSString *family = [[systemInfo[@"systemName"]
@@ -250,11 +252,17 @@ SentryCrashIntegration ()
     [deviceData setValue:systemInfo[@"freeMemory"] forKey:SentryDeviceContextFreeMemoryKey];
     [deviceData setValue:systemInfo[@"usableMemory"] forKey:@"usable_memory"];
     [deviceData setValue:systemInfo[@"memorySize"] forKey:@"memory_size"];
-    [deviceData setValue:systemInfo[@"storageSize"] forKey:@"storage_size"];
+    [deviceData setValue:systemInfo[@"totalStorageSize"] forKey:@"storage_size"];
+    [deviceData setValue:systemInfo[@"freeStorageSize"] forKey:@"free_storage"];
     [deviceData setValue:systemInfo[@"bootTime"] forKey:@"boot_time"];
 
     NSString *locale = [[NSLocale autoupdatingCurrentLocale] objectForKey:NSLocaleIdentifier];
     [deviceData setValue:locale forKey:LOCALE_KEY];
+
+#if SENTRY_HAS_UIDEVICE
+    [deviceData setValue:@(UIScreen.mainScreen.bounds.size.height) forKey:@"screen_height_pixels"];
+    [deviceData setValue:@(UIScreen.mainScreen.bounds.size.width) forKey:@"screen_width_pixels"];
+#endif
 
     [scope setContextValue:deviceData forKey:DEVICE_KEY];